Add files via upload
Decimal or hexadecimal number consists of odd as well as even number. Most of the times, it is required to check number is odd or even such as odd or even parity used in serisl communication. Hence, students will be able to check and count odd and even number in array using assembly language program.